+++ lineage/figures/figure7.py
@@ -1,10 +1,10 @@
 """
 File: figure7.py
 Purpose: Generates figure 7.
-Figure 7 analyzes heterogeneous (2 state), NOT censored,
+Figure 5 analyzes heterogeneous (2 state), censored (by both time and fate),
 single lineages (no more than one lineage per population)
-with similar proportions of cells in states but
-of varying distributions.
+with different proportions of cells in states by
+changing the values in the transition matrices.
 """
 import numpy as np
 
@@ -14,23 +14,23 @@ from .figureCommon import (
     commonAnalyze,
     figureMaker,
     pi,
+    E,
     max_desired_num_cells,
+    max_experiment_time,
     num_data_points,
-    state1,
 )
 from ..LineageTree import LineageTree
-from ..StateDistribution import StateDistribution
 
 
 def makeFigure():
     """
-    Makes figure 7.
+    Makes figure 5.
     """
 
     # Get list of axis objects
     ax, f = getSetup((7, 6), (2, 3))
 
-    figureMaker(ax, *accuracy(), xlabel="Wasserstein Divergence")
+    figureMaker(ax, *accuracy(), xlabel=r"Cells in State 0 [$\%$]")
 
     subplotLabel(ax)
 
@@ -40,21 +40,21 @@ def makeFigure():
 def accuracy():
     """
     Calculates accuracy and parameter estimation
-    over an increasing number of cells in a lineage for
+    over an similar number of cells in a lineage for
     a uncensored two-state model but differing state distribution.
-    We vary the distribution by
-    increasing the Wasserstein divergence between the two states.
+    We increase the proportion of cells in a lineage by
+    fixing the Transition matrix to be biased towards state 0.
     """
 
     # Creating a list of populations to analyze over
-    list_of_Es = [[StateDistribution(0.88, a, 1), state1] for a in np.logspace(1, 2, num_data_points, base=10)]
+    list_of_Ts = [np.array([[i, 1.0 - i], [i, 1.0 - i]]) for i in np.linspace(0.0, 1.0, num_data_points)]
     list_of_populations = []
-    for E in list_of_Es:
+    for T in list_of_Ts:
         population = []
 
-        population.append(LineageTree(pi, np.array([[0,1],[1,0]]), E, max_desired_num_cells))
+        population.append(LineageTree(pi, T, E, max_desired_num_cells, censor_condition=3, desired_experiment_time=max_experiment_time))
 
         # Adding populations into a holder for analysing
         list_of_populations.append(population)
 
-    return commonAnalyze(list_of_populations, xtype="wass")
+    return commonAnalyze(list_of_populations, xtype="prop")
